Extract decode_if_string and NXdata helpers into nexus.utils and nexus.nxdata
Move pure utility functions out of the 800-line nexus.py into proper
sub-modules with minimal dependencies:

- nexus/utils.py: decode_if_string (decodes HDF5 bytes/arrays to str)
- nexus/nxdata.py: chk_nxdata_axis*, logger_auxiliary_signal,
  print_default_plottable_header, entry_helper, nxdata_helper,
  signal_helper, find_attrib_axis_actual_dim_num,
  get_single_or_multiple_axes, axis_helper

nexus.py re-exports these via imports for backward compatibility.
annotation.py now imports decode_if_string, chk_nxdata_axis, and
get_default_plottable at module level (no more deferred local imports
for these three symbols). Tests updated accordingly.

Also fixes a pre-existing bug in chk_nxdata_axis_v2: `for i in len(axes)`
was corrected to `for i in range(len(axes))`.
