+++ src/pynxtools/nexus/utils.py
@@ -1,0 +1,82 @@
+#
+# Copyright The pynxtools Authors.
+#
+# This file is part of pynxtools.
+# See https://github.com/FAIRmat-NFDI/pynxtools for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""Low-level utility helpers for NeXus/HDF5 data decoding.
+
+This module intentionally has minimal dependencies so it can be imported
+safely from any layer of the pynxtools package without risk of circular
+imports.
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+import numpy as np
+
+from pynxtools.definitions.dev_tools.utils.nxdl_utils import decode_or_not
+
+
+def decode_if_string(
+    elem: Any, encoding: str = "utf-8", decode: bool = True
+) -> Any | None:
+    """
+    Decodes a numpy ndarray or list of byte objects or byte strings to strings.
+    If `decode` is False, returns the input value unchanged. Non-byte/str types
+    are returned without modification.
+
+    Args:
+        elem: A numpy ndarray, list, bytes, or any other object.
+        encoding: The encoding scheme to use. Default is "utf-8".
+        decode: A boolean flag indicating whether to perform decoding.
+
+    Returns:
+        A decoded string, a numpy array of decoded strings, a list of decoded strings,
+        or the input value if not decodable.
+        Returns None if the input is empty or invalid.
+
+    Raises:
+        ValueError: If decoding fails on bytes or numpy array elements.
+    """
+    # Early return if decoding is disabled
+    if not decode:
+        return elem
+
+    # Handle numpy arrays of bytes or strings
+    if isinstance(elem, np.ndarray):
+        if elem.size == 0:
+            return elem  # Return the empty array unchanged
+
+        # This only checks for null-terminated strings,
+        # may need to be updated in the future: https://api.h5py.org/h5t.html
+        if elem.dtype.kind == "S":  # Check if it's a bytes array (fixed-length strings)
+            decoded_array = np.vectorize(
+                lambda x: (
+                    x.decode(encoding).rstrip("\x00") if isinstance(x, bytes) else x
+                )
+            )(elem)
+            return decoded_array.astype(str)  # Ensure the dtype is str
+
+        # Handle mixed-type arrays
+        if elem.dtype == object:
+            decoded_array = np.vectorize(
+                lambda x: x.decode(encoding) if isinstance(x, bytes) else x
+            )(elem)
+            return decoded_array.astype(str)  # Ensure all elements are strings
+
+    return decode_or_not(elem, encoding, decode)
